feat(quality): schema merge + count_tokens + Retry-After + usage refresh + nextDateReset float
Phase 2 single-phase files. Cumulative quality fixes from the audit:

- internal/reqconv/schema_sanitize.go (Phase 2.5): mergeObjectBranches
  for anyOf/oneOf with multiple non-null branches — unions properties,
  intersects required across object branches. Heterogeneous fallback
  drops 'required' from the first branch so calls matching other
  variants aren't hard-rejected. Cuts 'lossy schema conversion' log
  spam (was firing 66× / 16 min on tools with union types).

- internal/messages/request.go (Phase 2.6): count_tokens now builds an
  Anthropic-shaped text representation (system + each message
  Role+Content + each tool Name+Description+Schema) rather than
  counting the Kiro wire payload (synthetic ack + ConversationState +
  ProfileARN + tool_specification wrapper, ~3 KiB/req overhead).
  Eliminates Claude Code / Cline over-trim issue.

- internal/kiroclient/backoff.go (Phase 2.7): effectiveDelay honors
  Retry-After on 429/5xx. Parses integer-seconds + HTTP-date forms;
  clamps to maxRetryAfter=30s; returns max(server, natural) so we're
  never more aggressive than the server requested but never wait
  longer than 30s if the server forgot to bound it.

- cmd/kiroxy/main.go + internal/kiroclient/usage.go (Phase 2.9 + 2.10):
  - Usage poller now refreshes-and-retries on 401/403, eliminating
    ~2031/2h log spam of 'bearer token invalid'. Shares RefreshConfig
    with chat path so a single refresh covers both via singleflight.
  - nextDateReset parser is float-tolerant. Upstream sometimes emits
    scientific notation (1.780272E9) which Go's int64 unmarshaler
    rejects; using *float64 + cast covers both forms.

- go.mod: golang.org/x/net v0.52.0 promoted to direct dep (required
  by Phase 2.1 typed http2 error classification, committed next).

+++ internal/kiroclient/usage.go
@@ -269,8 +269,17 @@ func parseUsageLimitsBody(body string) (*UsageLimits, error) {
 			TotalUsageLimit int64    `json:"totalUsageLimit"`
 			PercentUsed     *float64 `json:"percentUsed"`
 		} `json:"limits"`
-		NextDateReset  *int64 `json:"nextDateReset"`
-		DaysUntilReset *int   `json:"daysUntilReset"`
+		// NextDateReset is intentionally *float64 rather than *int64.
+		// Upstream emits this field in TWO inconsistent shapes: integer
+		// epoch seconds (1780272000) AND scientific-notation float
+		// (1.780272E9). Go's int64 unmarshaler rejects the latter with
+		// "cannot unmarshal JSON number 1.780272E9 into Go int64", which
+		// produces a hard parse error and discards the entire response.
+		// float64 accepts both forms and we round to int64 manually below.
+		// Float64 has 53 bits of mantissa, plenty for any plausible epoch
+		// second or millisecond value through the year 285,000.
+		NextDateReset  *float64 `json:"nextDateReset"`
+		DaysUntilReset *int     `json:"daysUntilReset"`
 	}
 	if err := json.Unmarshal([]byte(body), &raw); err != nil {
 		return nil, fmt.Errorf("getUsageLimits: parse json: %w", err)
@@ -304,7 +313,7 @@ func parseUsageLimitsBody(body string) (*UsageLimits, error) {
 		// Smithy conventions for this endpoint emit epoch seconds; guard
 		// against future drift by clamping implausible millisecond
 		// values.
-		ts := *raw.NextDateReset
+		ts := int64(*raw.NextDateReset)
 		if ts > 10_000_000_000 { // > Nov 2286 if seconds, so treat as ms
 			u.NextReset = time.UnixMilli(ts)
 		} else {
